requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
numpy==1.24.3
python-multipart==0.0.6
anyio==3.7.1
starlette==0.27.0
